fix: revert "refactor(core): abstract use bundle documents , reuse for variants (#13226)"
This reverts commit 6e71b377a99b82ed00a926dac47f4ce228ec214c.
