+++ packages/sanity/src/core/variants/tool/detail/__tests__/VariantDocumentsTable.test.tsx
@@ -1,5 +1,4 @@
 import {type SanityDocument} from '@sanity/client'
-import {type DocumentSystem} from '@sanity/types'
 import {render, screen, waitFor} from '@testing-library/react'
 import {userEvent} from '@testing-library/user-event'
 import {describe, expect, it, vi} from 'vitest'
@@ -40,13 +39,11 @@ const mockDocuments: SanityDocument[] = [
 ]
 
 describe('VariantDocumentsTable', () => {
-  const renderTable = async (documents: SanityDocument[] = mockDocuments, loading = false) => {
+  const renderTable = async (documents: SanityDocument[] = mockDocuments) => {
     const wrapper = await createTestProvider({
       resources: [variantsUsEnglishLocaleBundle],
     })
-    const result = render(<VariantDocumentsTable documents={documents} loading={loading} />, {
-      wrapper,
-    })
+    const result = render(<VariantDocumentsTable documents={documents} />, {wrapper})
     await screen.findByPlaceholderText('Search documents')
     return result
   }
@@ -55,13 +52,6 @@ describe('VariantDocumentsTable', () => {
     await renderTable([])
 
     expect(screen.getByText('No documents in this variant')).toBeInTheDocument()
-  })
-
-  it('shows loading skeleton rows while documents are loading', async () => {
-    await renderTable([], true)
-
-    expect(screen.getAllByTestId('table-row-skeleton')).toHaveLength(3)
-    expect(screen.queryByText('No documents in this variant')).not.toBeInTheDocument()
   })
 
   it('renders document rows with title, type, and edited columns', async () => {
@@ -94,58 +84,5 @@ describe('VariantDocumentsTable', () => {
 
     expect(screen.getByText('Second article')).toBeInTheDocument()
     expect(screen.queryByText('First article')).not.toBeInTheDocument()
-  })
-
-  it('groups documents by their document group ref', async () => {
-    const sharedGroupRef = 'article-group'
-    const groupedDocuments: SanityDocument[] = [
-      {
-        _id: 'versions.summer.article-a',
-        _type: 'article',
-        _rev: 'rev-1',
-        _createdAt: '2025-01-01T00:00:00Z',
-        _updatedAt: '2025-06-01T00:00:00Z',
-        title: 'Summer article',
-        _system: {
-          group: {_ref: sharedGroupRef, _weak: true},
-          bundleId: 'drafts',
-        } satisfies DocumentSystem,
-      },
-      {
-        _id: 'article-other',
-        _type: 'article',
-        _rev: 'rev-2',
-        _createdAt: '2025-01-02T00:00:00Z',
-        _updatedAt: '2025-06-02T00:00:00Z',
-        title: 'Other article',
-        _system: {
-          group: {_ref: 'other-group', _weak: true},
-          bundleId: 'drafts',
-        } satisfies DocumentSystem,
-      },
-      {
-        _id: 'versions.published.article-a',
-        _type: 'article',
-        _rev: 'rev-3',
-        _createdAt: '2025-01-03T00:00:00Z',
-        _updatedAt: '2025-06-03T00:00:00Z',
-        title: 'Published article',
-        _system: {
-          group: {_ref: sharedGroupRef, _weak: true},
-        } satisfies DocumentSystem,
-      },
-    ]
-
-    await renderTable(groupedDocuments)
-
-    await waitFor(() => {
-      expect(screen.getAllByTestId('table-row')).toHaveLength(3)
-    })
-
-    const rowTitles = screen.getAllByTestId('table-row').map((row) => row.textContent)
-
-    expect(rowTitles[0]).toContain('Published article')
-    expect(rowTitles[1]).toContain('Summer article')
-    expect(rowTitles[2]).toContain('Other article')
   })
 })
